Rename to jekyll-remote_csv
