+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in jekyll-everypolitician-education.gemspec
+# Specify your gem's dependencies in jekyll-remote_csv.gemspec
 gemspec
