+++ jekyll-remote_csv.gemspec
@@ -1,16 +1,16 @@
 # coding: utf-8
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'jekyll/everypolitician/education/version'
+require 'jekyll/remote_csv/version'
 
 Gem::Specification.new do |spec|
-  spec.name          = 'jekyll-everypolitician-education'
-  spec.version       = Jekyll::Everypolitician::Education::VERSION
+  spec.name          = 'jekyll-remote_csv'
+  spec.version       = Jekyll::RemoteCsv::VERSION
   spec.authors       = ['Chris Mytton']
   spec.email         = ['chrismytton@gmail.com']
 
-  spec.summary       = 'Add Education information to EveryPolitician data in Jekyll'
-  spec.homepage      = 'https://github.com/everypolitician/jekyll-everypolitician-education'
+  spec.summary       = 'Specify remote CSVs to be included in site.data'
+  spec.homepage      = 'https://github.com/everypolitician/jekyll-remote_csv'
   spec.license       = 'MIT'
 
   spec.files         = `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
@@ -19,7 +19,6 @@ Gem::Specification.new do |spec|
   spec.require_paths = ['lib']
 
   spec.add_dependency 'jekyll', '>= 3'
-  spec.add_dependency 'jekyll-everypolitician', '>= 0.4.0'
 
   spec.add_development_dependency 'bundler', '~> 1.10'
   spec.add_development_dependency 'rake', '~> 10.0'
