--- lib/jekyll/everypolitician/education/version.rb
@@ -1,7 +1,0 @@
-module Jekyll
-  module Everypolitician
-    module Education
-      VERSION = '0.1.0'
-    end
-  end
-end
