+++ lib/jekyll/remote_csv/version.rb
@@ -1,0 +1,5 @@
+module Jekyll
+  module RemoteCsv
+    VERSION = '0.1.0'
+  end
+end
