+++ test/fixtures/education.csv
@@ -7,4 +7,4 @@ id,name,organisation_name,role,subtitle,area,start_date,end_date
 19,Fani Munengami,,"Diploma in Marketing, Public Relations and Advertising",,,,
 22,Tendai Biti,University of Zimbabwe,Bachelor of Laws (LLB),,,,
 25,Fungai J Majome,University of Zimbabwe,Post Graduate Diploma,Women's Law,,,
-b61c6f06-586a-4c6a-aa83-d58b5b2d4e93,Fungai J Majome,University of Zimbabwe,Bachelor of Laws (LLB),,,,
+25,Fungai J Majome,University of Zimbabwe,Bachelor of Laws (LLB),,,,
