+++ test/test_helper.rb
@@ -1,5 +1,5 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'jekyll/everypolitician/education'
+require 'jekyll/remote_csv'
 
 Jekyll.logger.log_level = :error
